Move FeedCard out of MainActivity and resolve #34
